sharding: update child aggregator mode to latest smt changes
- join parent and child merkle paths in inclusion proof
- request id shard bits are now managed internally in the SMT
- set shardID type alias from int64 to int
- use shardID type alias instead of int

+++ internal/smt/smt.go
@@ -57,7 +57,7 @@ func NewChildSparseMerkleTree(algorithm api.HashAlgorithm, keyLength int, shardI
 	if shardID <= 1 {
 		panic("Shard ID must be positive and have at least 2 bits")
 	}
-	path := big.NewInt(shardID)
+	path := big.NewInt(int64(shardID))
 	if path.BitLen() > keyLength {
 		panic("Shard ID must be shorter than SMT key length")
 	}
